Add anagrams detection functionality and tests
This commit introduces a new functionality that allows detecting anagrams in an input list of words. A corresponding PowerShell script and a unit test file have also been added to ensure the correct functionality of the operation. The Anagram.ps1 script takes a subject word and a list of candidate words, and then identifies the subset that consists of anagrams of the subject word.