Add nursery rhyme recitation functionality
Implemented two new PowerShell functions, Get-Verse and Get-Rhyme. The Get-Verse function generates individual verses of 'This is the House that Jack Built'. The Get-Rhyme function uses Get-Verse to recite the nursery rhyme from a specified range of verses. Included a detailed README file and added extensive tests to validate functionality.