Add accumulate operation and tests in PowerShell
This commit introduces a new PowerShell function Get-Accumulation which takes a collection of numbers and an operation to perform on each element of the collection, returning a new collection with the results. Additionally, unit tests for this function have been added. A README has been added to provide instructions and details regarding the accumulate operation.